+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/internal/DiagnosticUtils.java
@@ -23,10 +23,6 @@ import java.util.stream.Stream;
 import org.eclipse.core.runtime.NullProgressMonitor;
 import org.eclipse.jdt.core.Flags;
 import org.eclipse.jdt.core.IAnnotation;
-import org.eclipse.jdt.core.dom.IAnnotationBinding;
-import org.eclipse.jdt.core.dom.IBinding;
-import org.eclipse.jdt.core.dom.IMethodBinding;
-import org.eclipse.jdt.core.dom.IVariableBinding;
 import org.eclipse.jdt.core.ICompilationUnit;
 import org.eclipse.jdt.core.IField;
 import org.eclipse.jdt.core.IImportContainer;
@@ -37,7 +33,6 @@ import org.eclipse.jdt.core.IMethod;
 import org.eclipse.jdt.core.IType;
 import org.eclipse.jdt.core.ITypeHierarchy;
 import org.eclipse.jdt.core.JavaModelException;
-import org.eclipse.jdt.core.dom.ITypeBinding;
 import org.eclipse.jdt.internal.corext.util.JavaModelUtil;
 import org.eclipse.lsp4jakarta.jdt.core.JakartaCorePlugin;
 
@@ -507,28 +502,6 @@ public class DiagnosticUtils {
     public static boolean isNegativePriorityValue(IAnnotation priorityAnnotation) throws JavaModelException {
         Number value = getAnnotationMemberValue(priorityAnnotation, "value", Number.class);
         return value != null && value.intValue() < 0;
-    }
-
-    /**
-     * Checks if the given binding (class, field, or method) has the specified annotation.
-     *
-     * @param binding The binding — may be an {@link ITypeBinding}, {@link IVariableBinding},
-     *            or {@link IMethodBinding}
-     * @param annotationFQN The fully qualified annotation name
-     * @return true if the annotation is present, false otherwise
-     */
-    public static boolean hasAnnotation(IBinding binding, String annotationFQN) {
-        IAnnotationBinding[] annotations;
-        if (binding instanceof ITypeBinding) {
-            annotations = ((ITypeBinding) binding).getAnnotations();
-        } else if (binding instanceof IVariableBinding) {
-            annotations = ((IVariableBinding) binding).getAnnotations();
-        } else if (binding instanceof IMethodBinding) {
-            annotations = ((IMethodBinding) binding).getAnnotations();
-        } else {
-            return false;
-        }
-        return Stream.of(annotations).anyMatch(annotation -> annotation.getAnnotationType().getQualifiedName().equals(annotationFQN));
     }
 
     /**
