+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/internal/ejb/RemoveSessionBeanAnnotationQuickFix.java
@@ -21,9 +21,9 @@ import org.eclipse.jdt.core.dom.IBinding;
 import org.eclipse.lsp4j.CodeAction;
 import org.eclipse.lsp4j.Diagnostic;
 import org.eclipse.lsp4jakarta.commons.codeaction.JakartaCodeActionId;
+import org.eclipse.lsp4jakarta.jdt.core.java.codeaction.CodeActionUtils;
 import org.eclipse.lsp4jakarta.jdt.core.java.codeaction.JavaCodeActionContext;
 import org.eclipse.lsp4jakarta.jdt.core.java.codeaction.RemoveAnnotationConflictQuickFix;
-import org.eclipse.lsp4jakarta.jdt.internal.DiagnosticUtils;
 
 /**
  * Removes the session bean annotation (@Stateless, @Stateful, or @Singleton)
@@ -61,7 +61,7 @@ public class RemoveSessionBeanAnnotationQuickFix extends RemoveAnnotationConflic
     protected void createCodeActions(Diagnostic diagnostic, JavaCodeActionContext context, IBinding parentType,
                                      List<CodeAction> codeActions) throws CoreException {
         // Only create a code action for each session bean annotation that is actually present
-        List<String> presentAnnotations = Arrays.stream(getAnnotations()).filter(annotation -> DiagnosticUtils.hasAnnotation(parentType, annotation)).collect(Collectors.toList());
+        List<String> presentAnnotations = Arrays.stream(getAnnotations()).filter(annotation -> CodeActionUtils.hasAnnotation(parentType, annotation)).collect(Collectors.toList());
         for (String annotation : presentAnnotations) {
             createCodeAction(diagnostic, context, parentType, codeActions, annotation);
         }
